Heredocs, no command in process and syntax error
Minishell was failing at handling more than one heredoc in one process.
Added new logic to pipex bonus so that it can handle a single process
with no command and just redirections. It doesn't work at the moment
because Alejandro has to return process->command=NULL in this case and
at the moment he's returning an empty string. I've also added the code
that handles the syntax error situation (exit_status = 258), but
Alejandro still has to modify this function.