Unsetting the first variable no longer segfaults
When deleting the first variable of the env t_list was causing minishell
to segfault. It was due to the fact that I was losing the reference to
the outter env_list t_list (the one declared in the 'main' function)
because I declared the fields 'env' and 'export_list' of the t_prompt
struct as single pointers and doing prompt->env = *env and
prompt->export_list = *export_list. Changing both fields to t_list **
and doing prompt->env = env and prompt->export_list = export_list solved
the issue.